feat: enhance number card functionality with dynamic filters and percentage display
