+++ apps/app/lib/number_card/cubit/number_card_state.dart
@@ -6,6 +6,7 @@ class NumberCardState extends Equatable {
     this.title,
     this.value,
     this.percent,
+    this.showPercentage,
   });
 
   // initial state with default values
@@ -15,17 +16,20 @@ class NumberCardState extends Equatable {
   final String? title;
   final String? value;
   final String? percent;
+  final bool? showPercentage;
 
   NumberCardState copyWith({
     String? id,
     String? title,
     String? value,
     String? percent,
+    bool? showPercentage,
   }) => NumberCardState(
     id: id ?? this.id,
     title: title ?? this.title,
     value: value ?? this.value,
     percent: percent ?? this.percent,
+    showPercentage: showPercentage ?? this.showPercentage,
   );
 
   @override
@@ -34,5 +38,6 @@ class NumberCardState extends Equatable {
     title,
     value,
     percent,
+    showPercentage,
   ];
 }
