+++ apps/app/lib/number_card/view/view.dart
@@ -1,1 +1,2 @@
 export './number_card_page.dart';
+export './number_card_view.dart';
